[FLINK-33905][core] Introduce the interface and default implementation of JobInfo and TaskInfo
