+++ flink-core/src/main/java/org/apache/flink/api/common/operators/CollectionExecutor.java
@@ -25,6 +25,7 @@ import org.apache.flink.api.common.JobExecutionResult;
 import org.apache.flink.api.common.JobID;
 import org.apache.flink.api.common.Plan;
 import org.apache.flink.api.common.TaskInfo;
+import org.apache.flink.api.common.TaskInfoImpl;
 import org.apache.flink.api.common.accumulators.Accumulator;
 import org.apache.flink.api.common.accumulators.AccumulatorHelper;
 import org.apache.flink.api.common.aggregators.Aggregator;
@@ -183,7 +184,7 @@ public class CollectionExecutor {
         GenericDataSinkBase<IN> typedSink = (GenericDataSinkBase<IN>) sink;
 
         // build the runtime context and compute broadcast variables, if necessary
-        TaskInfo taskInfo = new TaskInfo(typedSink.getName(), 1, 0, 1, 0);
+        TaskInfo taskInfo = new TaskInfoImpl(typedSink.getName(), 1, 0, 1, 0);
         RuntimeUDFContext ctx;
 
         if (RichOutputFormat.class.isAssignableFrom(
@@ -222,7 +223,7 @@ public class CollectionExecutor {
         @SuppressWarnings("unchecked")
         GenericDataSourceBase<OUT, ?> typedSource = (GenericDataSourceBase<OUT, ?>) source;
         // build the runtime context and compute broadcast variables, if necessary
-        TaskInfo taskInfo = new TaskInfo(typedSource.getName(), 1, 0, 1, 0);
+        TaskInfo taskInfo = new TaskInfoImpl(typedSource.getName(), 1, 0, 1, 0);
 
         RuntimeUDFContext ctx;
 
@@ -250,7 +251,7 @@ public class CollectionExecutor {
         SingleInputOperator<IN, OUT, ?> typedOp = (SingleInputOperator<IN, OUT, ?>) operator;
 
         // build the runtime context and compute broadcast variables, if necessary
-        TaskInfo taskInfo = new TaskInfo(typedOp.getName(), 1, 0, 1, 0);
+        TaskInfo taskInfo = new TaskInfoImpl(typedOp.getName(), 1, 0, 1, 0);
         RuntimeUDFContext ctx;
 
         if (RichFunction.class.isAssignableFrom(typedOp.getUserCodeWrapper().getUserCodeClass())) {
@@ -293,7 +294,7 @@ public class CollectionExecutor {
                 (DualInputOperator<IN1, IN2, OUT, ?>) operator;
 
         // build the runtime context and compute broadcast variables, if necessary
-        TaskInfo taskInfo = new TaskInfo(typedOp.getName(), 1, 0, 1, 0);
+        TaskInfo taskInfo = new TaskInfoImpl(typedOp.getName(), 1, 0, 1, 0);
         RuntimeUDFContext ctx;
 
         if (RichFunction.class.isAssignableFrom(typedOp.getUserCodeWrapper().getUserCodeClass())) {
