+++ flink-core/src/test/java/org/apache/flink/api/common/io/RichInputFormatTest.java
@@ -20,6 +20,7 @@ package org.apache.flink.api.common.io;
 
 import org.apache.flink.api.common.ExecutionConfig;
 import org.apache.flink.api.common.TaskInfo;
+import org.apache.flink.api.common.TaskInfoImpl;
 import org.apache.flink.api.common.accumulators.Accumulator;
 import org.apache.flink.api.common.functions.util.RuntimeUDFContext;
 import org.apache.flink.core.fs.Path;
@@ -39,7 +40,7 @@ public class RichInputFormatTest {
     @Test
     public void testCheckRuntimeContextAccess() {
         final SerializedInputFormat<Value> inputFormat = new SerializedInputFormat<Value>();
-        final TaskInfo taskInfo = new TaskInfo("test name", 3, 1, 3, 0);
+        final TaskInfo taskInfo = new TaskInfoImpl("test name", 3, 1, 3, 0);
         inputFormat.setRuntimeContext(
                 new RuntimeUDFContext(
                         taskInfo,
