+++ flink-core/src/test/java/org/apache/flink/api/common/operators/base/FlatMapOperatorCollectionTest.java
@@ -20,6 +20,7 @@ package org.apache.flink.api.common.operators.base;
 
 import org.apache.flink.api.common.ExecutionConfig;
 import org.apache.flink.api.common.TaskInfo;
+import org.apache.flink.api.common.TaskInfoImpl;
 import org.apache.flink.api.common.accumulators.Accumulator;
 import org.apache.flink.api.common.functions.FlatMapFunction;
 import org.apache.flink.api.common.functions.OpenContext;
@@ -79,7 +80,7 @@ public class FlatMapOperatorCollectionTest implements Serializable {
         } else {
             executionConfig.enableObjectReuse();
         }
-        final TaskInfo taskInfo = new TaskInfo("Test UDF", 4, 0, 4, 0);
+        final TaskInfo taskInfo = new TaskInfoImpl("Test UDF", 4, 0, 4, 0);
         // run on collections
         final List<String> result =
                 getTestFlatMapOperator(udf)
