+++ flink-core/src/test/java/org/apache/flink/api/common/operators/base/OuterJoinOperatorBaseTest.java
@@ -20,6 +20,7 @@ package org.apache.flink.api.common.operators.base;
 
 import org.apache.flink.api.common.ExecutionConfig;
 import org.apache.flink.api.common.TaskInfo;
+import org.apache.flink.api.common.TaskInfoImpl;
 import org.apache.flink.api.common.accumulators.Accumulator;
 import org.apache.flink.api.common.functions.FlatJoinFunction;
 import org.apache.flink.api.common.functions.OpenContext;
@@ -79,7 +80,7 @@ public class OuterJoinOperatorBaseTest implements Serializable {
         executionConfig = new ExecutionConfig();
 
         String taskName = "Test rich outer join function";
-        TaskInfo taskInfo = new TaskInfo(taskName, 1, 0, 1, 0);
+        TaskInfo taskInfo = new TaskInfoImpl(taskName, 1, 0, 1, 0);
         HashMap<String, Accumulator<?, ?>> accumulatorMap = new HashMap<>();
         HashMap<String, Future<Path>> cpTasks = new HashMap<>();
